Add machinery for tracking ProviderConfig usage
This commit introduces a new type and controller that can be used to track the
managed resources that are using a ProviderConfig. This allows us to block the
deletion of a ProviderConfig while resources are still using it.

Note the new type is primarily useful when a provider introduces several
different kinds of managed resource that may all use the same kind of
ProviderConfig. This is not the case in the Helm provider, which has only one
resource (Release). We introduce the new type in this commit in order to comply
with the pattern that will be used by all providers in future.

Signed-off-by: Nic Cope <negz@rk0n.org>