Avoid up-to-date check when a release is being deleted
When a release is being deleted there's a high likelihood that its dependencies
(i.e. the secrets and configmaps it uses to set values) are being deleted too.
When this is the case it's impossible to determine whether the release is up to
date, and thus it blocks forever.

The managed resource reconciler doesn't actually need to know whether the
release is up to date when it's being deleted - it just needs to know whether
the release exists or not.

Signed-off-by: Nic Cope <negz@rk0n.org>