Support GKE OAuth Authorization
This commit adds an optional set of 'identity' credentials to the Helm provider
config. Currently only the GoogleApplicationCredentials type of identity is
supported. If supplied, the kubeconfig (fetched from the regular 'credentials')
is used to create an API Server REST client, which is wrapped such that it will
use a Google Application Credentials file to fetch OAuth tokens which are then
used as the Bearer token to authorize requests.

Signed-off-by: Nic Cope <negz@rk0n.org>